# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Via-Alias API",
    "description": "API for [Via-Alias](https://github.com/ngundlach/via-alias). This API can be used to create and manage redirects as well as users.\n        <br><br>**Authentication:**<br>\n        Most Endpoints require a Bearer token obtained from `POST /api/auth/login`.\n        Include it in the `Authorization` header as `Bearer <token>`.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.11"
  },
  "servers": [
    {
      "url": "https://example.com",
      "description": "Example:"
    }
  ],
  "paths": {
    "/api/admin/redirects": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get all redirects",
        "description": "Returns a list of all currently created redirects. Requires authentication. Pass a JWT as a bearer token in the `Authorization` header.",
        "operationId": "get_all_redirects",
        "responses": {
          "200": {
            "description": "Success. Returns a list of all currently created redirects",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FullRedirectListDTO"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized. No valid access token."
          },
          "403": {
            "description": "Forbidden. User authorized but doesn't have permission."
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/admin/redirects/{id}": {
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "Delete redirect via id",
        "description": "Deletes a redirect via its id. Requires authentication. Pass a JWT as a bearer token in the `Authorization` header.",
        "operationId": "delete_redirect_by_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The redirect id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content. Deletes a redirect"
          },
          "401": {
            "description": "Unauthorized. No valid access token."
          },
          "403": {
            "description": "Forbidden. User authorized but doesn't have permission."
          },
          "404": {
            "description": "Not Found. Redirect doesn't exist."
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/admin/reg_token": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Generate user registration token",
        "description": "Requests a user registration token. This token is only valid for one-time use and has an expiration time in seconds. Requires authentication. Pass a JWT as a bearer token in the `Authorization` header.",
        "operationId": "reg_token_request",
        "responses": {
          "200": {
            "description": "Success. Returns valid registration token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRegistrationTokenDTO"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized. No valid access token."
          },
          "403": {
            "description": "Forbidden. User authorized but doesn't have permission."
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/admin/users": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get all users",
        "description": "Returns a list of all users. Requires authentication. Pass a JWT as a bearer token in the `Authorization` header.",
        "operationId": "get_all_users_data_admin",
        "responses": {
          "200": {
            "description": "Ok. Returns list of persisted userdata.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserListDTO"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized. No valid access token."
          },
          "403": {
            "description": "Forbidden. User authorized but doesn't have permission."
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/admin/users/{id}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get user data",
        "description": "Returns data about a specific user. Requires authentication. Pass a JWT as a bearer token in the `Authorization` header.",
        "operationId": "get_user_data_admin",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The user id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ok. Returns persisted userdata.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized. No valid access token."
          },
          "403": {
            "description": "Forbidden. User authorized but doesn't have permission."
          },
          "404": {
            "description": "Not Found. User doesn't exist."
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "Delete a user",
        "description": "Deletes a user via its uiserid. This will also delete all of the users registered redirects. Requires authentication. Pass a JWT as a bearer token in the `Authorization` header.",
        "operationId": "delete_user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The user id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ok. Returns userid and amount of deleted redirects",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletedUserDTO"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized. No valid access token."
          },
          "403": {
            "description": "Forbidden. User authorized but doesn't have permission."
          },
          "404": {
            "description": "Not Found. User doesn't exist."
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login",
        "description": "Authenticates a user with their credentials and returns a signed JWT access token on success.\n    The token should be included in subsequent requests as a Bearer token in the `Authorization` header.",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCredentialsDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "User authenticated. Returns valid JWT access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserTokenDTO"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized. Invalid username or password"
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
      "DeletedUserDTO": {
        "type": "object",
        "title": "DeletedUser",
        "required": [
          "user_id",
          "deleted"
        ],
        "properties": {
          "deleted": {
            "$ref": "#/components/schemas/DeletedUserResourceDTO"
          },
          "user_id": {
            "type": "string",
            "examples": [
              "d64bcaad-8d86-48d2-b1f3-f1c03ac30fa3"
            ]
          }
        }
      },
      "DeletedUserResourceDTO": {
        "type": "object",
        "title": "DeletedUserResource",
        "required": [
          "redirects"
        ],
        "properties": {
          "redirects": {
            "type": "integer",
            "format": "int64",
            "examples": [
              5
            ],
            "minimum": 0
          }
        }
      },
      "FullRedirectListDTO": {
        "type": "object",
        "title": "FullRedirectList",
        "required": [
          "redirects"
        ],
        "properties": {
          "redirects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Redirect"
            }
          }
        }
      },
...
      "Redirect": {
        "type": "object",
        "title": "RedirectData",
        "required": [
          "id",
          "alias",
          "url",
          "owner"
        ],
        "properties": {
          "alias": {
            "type": "string",
            "examples": [
              "gh"
            ]
          },
          "id": {
            "type": "string",
            "examples": [
              "ea07b388-0da5-4640-b30d-2f90467a612c"
            ]
          },
          "owner": {
            "type": "string",
            "examples": [
              "7484bf63-0c9a-41af-884e-e0fea7f0bb8e"
            ]
          },
          "url": {
            "type": "string",
            "examples": [
              "http://www.github.com"
            ]
          }
        }
      },
...
      "UserCredentialsDTO": {
        "type": "object",
        "title": "UserCredentials",
        "required": [
          "name",
          "pw"
        ],
        "properties": {
          "name": {
            "type": "string",
            "examples": [
              "luke"
            ]
          },
          "pw": {
            "type": "string",
            "examples": [
              "superjedimeister1337"
            ]
          }
        }
      },
      "UserDTO": {
        "type": "object",
        "title": "UserData",
        "required": [
          "id",
          "name",
          "is_admin"
        ],
        "properties": {
          "id": {
            "type": "string",
            "examples": [
              "d64bcaad-8d86-48d2-b1f3-f1c03ac30fa3"
            ]
          },
          "is_admin": {
            "type": "boolean",
            "examples": [
              false
            ]
          },
          "name": {
            "type": "string",
            "examples": [
              "luke"
            ]
          }
        }
      },
      "UserListDTO": {
        "type": "object",
        "title": "UserList",
        "required": [
          "users"
        ],
        "properties": {
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserDTO"
            }
          }
        }
      },
...
      "UserRegistrationTokenDTO": {
        "type": "object",
        "title": "UserRegistrationToken",
        "required": [
          "registration_token",
          "token_ttl"
        ],
        "properties": {
          "registration_token": {
            "type": "string",
            "examples": [
              "85e83a5a-4f49-4ea7-9df9-93c2c2cc9b8f"
            ]
          },
          "token_ttl": {
            "type": "integer",
            "format": "int64",
            "examples": [
              1800
            ],
            "minimum": 0
          }
        }
      },
      "UserTokenDTO": {
        "type": "object",
        "title": "UserAccessToken",
        "required": [
          "access_token",
          "token_type",
          "expires_in"
        ],
        "properties": {
          "access_token": {
            "type": "string",
            "examples": [
              "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzUxMiJ9.eyJ1c2VyX2lkIjoiMjE1ZDU3YzAtMDM1My00Nzc2LWFjNTYtZDBiZWY5YTUxZTJjIiwiaXNfYWRtaW4iOmZhbHNlLCJleHAiOjE3NzI4OTczOTMsImp0aSI6IjRiZTcyNTUwLTZlYTUtNGQ3NS1iM2Q4LTdiNzJiYTEwZjE1MiJ9.qwy7zLn611SkZzI5mFJPwtRGjvjD0xmprSoMUII7xjcxRPjAbTEKH9gFIewYRGdwtcg0I-EhnttYMNZkqlmZNQ"
            ]
          },
          "expires_in": {
            "type": "integer",
            "format": "int64",
            "examples": [
              900
            ],
            "minimum": 0
          },
          "token_type": {
            "type": "string",
            "examples": [
              "Bearer"
            ]
          }
        }
      }
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}